# fix performance issue in torch.sparse.mm reduce mode